[refactor] cannot find class: UserForAuth 관련 오류 수정
